requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
anyio @ file:///tmp/build/80754af9/anyio_1644481698350/work/dist
argon2-cffi @ file:///opt/conda/conda-bld/argon2-cffi_1645000214183/work
argon2-cffi-bindings @ file:///tmp/build/80754af9/argon2-cffi-bindings_1644569684262/work
arrow==1.2.3
asttokens @ file:///opt/conda/conda-bld/asttokens_1646925590279/work
async-timeout==4.0.2
attrs==22.2.0
autopep8 @ file:///opt/conda/conda-bld/autopep8_1650463822033/work
Babel @ file:///croot/babel_1671781930836/work
backcall @ file:///home/ktietz/src/ci/backcall_1611930011877/work
beautifulsoup4 @ file:///opt/conda/conda-bld/beautifulsoup4_1650462163268/work
bleach @ file:///opt/conda/conda-bld/bleach_1641577558959/work
blessed==1.20.0
blis==0.2.4
brotlipy==0.7.0
certifi @ file:///croot/certifi_1671487769961/work/certifi
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer==3.1.0
click==8.1.3
comm @ file:///croot/comm_1671231121260/work
contourpy==1.0.7
croniter==1.3.8
cryptography @ file:///croot/cryptography_1677533068310/work
cycler==0.11.0
cymem==2.0.7
dateutils==0.6.12
debugpy @ file:///tmp/build/80754af9/debugpy_1637091796427/work
decorator @ file:///opt/conda/conda-bld/decorator_1643638310831/work
deepdiff==6.3.0
defusedxml @ file:///tmp/build/80754af9/defusedxml_1615228127516/work
dnspython==2.3.0
email-validator==1.3.1
en-core-web-sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.1.0/en_core_web_sm-2.1.0.tar.gz
entrypoints @ file:///tmp/build/80754af9/entrypoints_1649926445639/work
executing @ file:///opt/conda/conda-bld/executing_1646925071911/work
fastapi==0.88.0
fastjsonschema @ file:///opt/conda/conda-bld/python-fastjsonschema_1661371079312/work
filelock==3.9.0
flit_core @ file:///opt/conda/conda-bld/flit-core_1644941570762/work/source/flit_core
fonttools==4.39.0
frozenlist==1.3.3
fsspec==2023.3.0
h11==0.14.0
httpcore==0.16.3
httptools==0.5.0
httpx==0.23.3
huggingface-hub==0.11.1
hydra-core==1.3.2
idna @ file:///croot/idna_1666125576474/work
importlib-metadata @ file:///tmp/build/80754af9/importlib-metadata_1648562408398/work
importlib-resources @ file:///tmp/build/80754af9/importlib_resources_1625135880749/work
inquirer==3.1.3
ipykernel @ file:///croot/ipykernel_1671488378391/work
ipython @ file:///croot/ipython_1676582224036/work
ipython-genutils @ file:///tmp/build/80754af9/ipython_genutils_1606773439826/work
ipywidgets @ file:///tmp/build/80754af9/ipywidgets_1634143127070/work
itsdangerous==2.1.2
jedi @ file:///tmp/build/80754af9/jedi_1644315233700/work
Jinja2 @ file:///croot/jinja2_1666908132255/work
joblib==1.2.0
json5 @ file:///tmp/build/80754af9/json5_1624432770122/work
jsonschema @ file:///croot/jsonschema_1676558650973/work
jupyter @ file:///tmp/abs_33h4eoipez/croots/recipe/jupyter_1659349046347/work
jupyter-console @ file:///croot/jupyter_console_1677674649642/work
jupyter-server @ file:///croot/jupyter_server_1671707632269/work
jupyter_client @ file:///croot/jupyter_client_1676329080601/work
jupyter_core @ file:///croot/jupyter_core_1676538566912/work
jupyterlab @ file:///croot/jupyterlab_1675354114448/work
jupyterlab-pygments @ file:///tmp/build/80754af9/jupyterlab_pygments_1601490720602/work
jupyterlab-widgets @ file:///tmp/build/80754af9/jupyterlab_widgets_1609884341231/work
jupyterlab_server @ file:///croot/jupyterlab_server_1677143054853/work
kiwisolver==1.4.4
lightning==2.0.0
lightning-cloud==0.5.32
lightning-utilities==0.8.0
lxml @ file:///opt/conda/conda-bld/lxml_1657545139709/work
markdown-it-py==2.2.0
MarkupSafe @ file:///opt/conda/conda-bld/markupsafe_1654597864307/work
matplotlib==3.7.1
matplotlib-inline @ file:///opt/conda/conda-bld/matplotlib-inline_1662014470464/work
mdurl==0.1.2
mistune==0.8.4
multidict==6.0.4
murmurhash==1.0.9
nbclassic @ file:///croot/nbclassic_1676902904845/work
nbclient @ file:///tmp/build/80754af9/nbclient_1650308366712/work
nbconvert @ file:///croot/nbconvert_1668450669124/work
nbformat @ file:///croot/nbformat_1670352325207/work
nest-asyncio @ file:///croot/nest-asyncio_1672387112409/work
notebook @ file:///croot/notebook_1668179881751/work
notebook_shim @ file:///croot/notebook-shim_1668160579331/work
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
omegaconf==2.3.0
ordered-set==4.1.0
orjson==3.8.8
packaging==23.0
pandocfilters @ file:///opt/conda/conda-bld/pandocfilters_1643405455980/work
parso @ file:///opt/conda/conda-bld/parso_1641458642106/work
pexpect @ file:///tmp/build/80754af9/pexpect_1605563209008/work
pickleshare @ file:///tmp/build/80754af9/pickleshare_1606932040724/work
Pillow==9.4.0
pkgutil_resolve_name @ file:///opt/conda/conda-bld/pkgutil-resolve-name_1661463321198/work
plac==0.9.6
platformdirs @ file:///opt/conda/conda-bld/platformdirs_1662711380096/work
ply==3.11
preshed==2.0.1
prometheus-client @ file:///tmp/abs_d3zeliano1/croots/recipe/prometheus_client_1659455100375/work
prompt-toolkit @ file:///croot/prompt-toolkit_1672387306916/work
psutil @ file:///opt/conda/conda-bld/psutil_1656431268089/work
ptyprocess @ file:///tmp/build/80754af9/ptyprocess_1609355006118/work/dist/ptyprocess-0.7.0-py2.py3-none-any.whl
pure-eval @ file:///opt/conda/conda-bld/pure_eval_1646925070566/work
pycodestyle @ file:///croot/pycodestyle_1674267221883/work
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.7
Pygments==2.14.0
pyhocon==0.3.51
PyJWT==2.6.0
pyOpenSSL @ file:///croot/pyopenssl_1677607685877/work
pyparsing==3.0.9
PyQt5-sip==12.11.0
pyrsistent==0.19.3
PySocks @ file:///tmp/build/80754af9/pysocks_1605305779399/work
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
python-dotenv==1.0.0
python-editor==1.0.4
python-multipart==0.0.6
pytorch-lightning==2.0.0
pytz @ file:///croot/pytz_1671697431263/work
PyYAML==6.0
pyzmq @ file:///opt/conda/conda-bld/pyzmq_1657724186960/work
qtconsole @ file:///croot/qtconsole_1674008428467/work
QtPy @ file:///opt/conda/conda-bld/qtpy_1662014892439/work
readchar==4.0.5
regex==2022.10.31
requests==2.28.2
rfc3986==1.5.0
rich==13.3.3
scikit-learn==1.2.2
scipy==1.10.1
Send2Trash @ file:///tmp/build/80754af9/send2trash_1632406701022/work
sip @ file:///tmp/abs_44cd77b_pu/croots/recipe/sip_1659012365470/work
six @ file:///tmp/build/80754af9/six_1644875935023/work
sniffio @ file:///tmp/build/80754af9/sniffio_1614030475067/work
soupsieve @ file:///croot/soupsieve_1666296392845/work
spacy==2.1.4
srsly==1.0.6
stack-data @ file:///opt/conda/conda-bld/stack_data_1646927590127/work
starlette==0.22.0
starsessions==1.3.0
terminado @ file:///croot/terminado_1671751832461/work
thinc==7.0.8
threadpoolctl==3.1.0
tinycss2 @ file:///croot/tinycss2_1668168815555/work
tokenizers==0.13.2
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
tomli @ file:///opt/conda/conda-bld/tomli_1657175507142/work
torch==1.13.1
torchmetrics==0.11.4
tornado @ file:///opt/conda/conda-bld/tornado_1662061693373/work
tqdm==4.65.0
traitlets @ file:///croot/traitlets_1671143879854/work
transformers==4.26.1
typing_extensions==4.5.0
ujson==5.7.0
urllib3==1.26.15
uvicorn==0.21.1
uvloop==0.17.0
wasabi==0.10.1
watchfiles==0.18.1
wcwidth @ file:///Users/ktietz/demo/mc3/conda-bld/wcwidth_1629357192024/work
webencodings==0.5.1
websocket-client @ file:///tmp/build/80754af9/websocket-client_1614804261064/work
websockets==10.4
widgetsnbextension @ file:///tmp/build/80754af9/widgetsnbextension_1645009353553/work
yarl==1.8.2
zipp @ file:///croot/zipp_1672387121353/work
